Add files via upload
Readme:
Last updated 01.20.2025
AHKControl for AutoHotkey Version 2 by thehafenator. This is a 'port' from Lexicos' (The AutoHotkey Developer) AHKControl written in V1, and contains much of the functionality. Though not a true port, I was able to bring most of the functionality. 
See the V1 script on github here: https://github.com/Lexikos/AHKControl.ahk

Function of this V2 code:
1. View all AHK scripts, V1 or V2, in a single tray icon.
2. View each AHK script's original TrayMenu through the menu creation.
3. Send messages to reload, edit, suspend, pause, or edit individual scripts.
4. The 'Edit' command will show a list of all of the #include files in an AHK script, allowing each to be accessed.
5. An editor path (the variable is 'editorpath') can be specified to allow the scripts to be opened with 'edit' in an editor of choice. If the editor path does not work, it will default to opening with notepad. 
6. Pause/Suspend checkmarks. The script will show checkmarks under each submenu if the respective script is paused or suspended. It does this in three ways:
a. On startup or reload, it runs a wmi query to check the suspended/paused status of each script. 
b. By default, the script updates every 15 seconds
c. If a script is paused through an AHKControl submenu, it keeps an internal map called scriptstates(). This keeps the checkmarks more accurate until the timer refreshes sets the script state again.
Though not perfect, the only way that the pause/suspend checkmarks will be inaccurate will be if the user suspends a script through that script's own icon. However, it will only remain out of sync for 15 seconds or if the user reloads AHKControl, whichever comes first.
7. Apply a different icon for light and dark theme in Windows 10 and 11. The script reads a registry value to determine if it is in dark theme, and will try to set the theme depending on which value it reads, and have a back in case it doesn't work. If all do not work, the script will proceed with the default AHK icon.
8. Dark Mode Win 32 menus. With contextcolor(), which was provided on a forum response post by Lexicos to another user.


Funcationality that I would like to add, but am not sure how to implement:
1. Adding .exe/compiled AHK scripts into the GetRunningAHKScripts(). I haven't been able to figure this one out. If anyone wants to help, that would be great. 

Differences that that others may be able to tackle:
1. Grab icons from the scripts themselves. The V1 script was able to find and identify the icons within each script itself, even when #Notrayicon was used. I personally prefer the no icon look and think it looks cleaner.
2. The Scripts were numbered on Lexicos' version within A_TrayMenu, though I haven't been able how to achieve this and keep the paths clean throughout the script. 
I also prefer this look without the numbering.